CC-39568 Extract shared TokenHelper compat resolver into a trait
