+++ Spryker/Sniffs/AbstractSniffs/AbstractSprykerSniff.php
@@ -16,10 +16,12 @@ use SlevomatCodingStandard\Helpers\EmptyFileException;
 use SlevomatCodingStandard\Helpers\NamespaceHelper;
 use SlevomatCodingStandard\Helpers\TokenHelper;
 use Spryker\Traits\BasicsTrait;
+use Spryker\Traits\TokenHelperCompatTrait;
 
 abstract class AbstractSprykerSniff implements Sniff
 {
     use BasicsTrait;
+    use TokenHelperCompatTrait;
 
     /**
      * @var string
@@ -183,36 +185,6 @@ abstract class AbstractSprykerSniff implements Sniff
             $phpCsFile,
             $prevIndex,
         );
-    }
-
-    /**
-     * Resolves a Slevomat `TokenHelper` token-code list by name, tolerating its move from a
-     * public static property (Slevomat < 8.16) to a class constant (Slevomat >= 8.16). The
-     * name is resolved dynamically so neither form appears as a static literal that would be
-     * rejected by static analysis against whichever Slevomat version is installed.
-     *
-     * @param string $constantName
-     * @param string $propertyName
-     *
-     * @return array<int|string>
-     */
-    protected function resolveTokenHelperCodes(string $constantName, string $propertyName): array
-    {
-        $qualifiedConstantName = TokenHelper::class . '::' . $constantName;
-        $codes = defined($qualifiedConstantName)
-            ? constant($qualifiedConstantName)
-            : TokenHelper::${$propertyName};
-
-        $result = [];
-        if (is_array($codes)) {
-            foreach ($codes as $code) {
-                if (is_int($code) || is_string($code)) {
-                    $result[] = $code;
-                }
-            }
-        }
-
-        return $result;
     }
 
     /**
