+++ Spryker/Sniffs/Commenting/DisallowArrayTypeHintSyntaxSniff.php
@@ -30,6 +30,7 @@ use SlevomatCodingStandard\Helpers\NamespaceHelper;
 use SlevomatCodingStandard\Helpers\SniffSettingsHelper;
 use SlevomatCodingStandard\Helpers\TokenHelper;
 use SlevomatCodingStandard\Helpers\TypeHintHelper;
+use Spryker\Traits\TokenHelperCompatTrait;
 
 /**
  * Fixed version of Slevomatic, touching collection objects the right way.
@@ -38,6 +39,8 @@ use SlevomatCodingStandard\Helpers\TypeHintHelper;
  */
 class DisallowArrayTypeHintSyntaxSniff implements Sniff
 {
+    use TokenHelperCompatTrait;
+
     /**
      * @var string
      */
@@ -354,36 +357,6 @@ class DisallowArrayTypeHintSyntaxSniff implements Sniff
         }
 
         return null;
-    }
-
-    /**
-     * Resolves a Slevomat `TokenHelper` token-code list by name, tolerating its move from a
-     * public static property (Slevomat < 8.16) to a class constant (Slevomat >= 8.16). The
-     * name is resolved dynamically so neither form appears as a static literal that would be
-     * rejected by static analysis against whichever Slevomat version is installed.
-     *
-     * @param string $constantName
-     * @param string $propertyName
-     *
-     * @return array<int|string>
-     */
-    protected function resolveTokenHelperCodes(string $constantName, string $propertyName): array
-    {
-        $qualifiedConstantName = TokenHelper::class . '::' . $constantName;
-        $codes = defined($qualifiedConstantName)
-            ? constant($qualifiedConstantName)
-            : TokenHelper::${$propertyName};
-
-        $result = [];
-        if (is_array($codes)) {
-            foreach ($codes as $code) {
-                if (is_int($code) || is_string($code)) {
-                    $result[] = $code;
-                }
-            }
-        }
-
-        return $result;
     }
 
     /**
